+++ internal/mode/export/dotenv/doc.go
@@ -1,0 +1,3 @@
+// Package dotenv is a Ferrite mode that exports the environment variables to a
+// file suitable use as a .env file.
+package dotenv